agents/shopgenie-agent/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
httpx==0.27.2
python-dotenv==1.0.1
firebase-admin==6.5.0
